[REFACTOR] pointhistory - 구조 갱신
