+++ be/src/main/java/com/dd/blog/domain/admin/controller/AdminUserController.java
@@ -6,8 +6,8 @@ import com.dd.blog.domain.admin.dto.AdminUserRoleUpdateRequestDto;
 import com.dd.blog.domain.admin.dto.AdminUserStatusUpdateRequestDto;
 import com.dd.blog.domain.admin.dto.UserInfoResponseDto;
 import com.dd.blog.domain.admin.service.AdminUserService;
-import com.dd.blog.domain.point.dto.PointHistoryResponseDto;
-import com.dd.blog.domain.point.service.PointHistoryService;
+import com.dd.blog.domain.point.point.dto.PointHistoryResponseDto;
+import com.dd.blog.domain.point.point.service.PointHistoryService;
 import com.dd.blog.domain.user.user.entity.UserRole;
 import com.dd.blog.domain.user.user.entity.UserStatus;
 import com.dd.blog.global.security.SecurityUser;
